Merge main at 7792e94e65 and resolve executor, state, and test conflicts. Preserve typed outcomes while adopting the updated BAL and database APIs. Carry superseded write sets through operational results and release them during apply-loop teardown.

Validation: make lint; make erigon integration; affected-package unit tests; race tests for common/errors, execution/exec, execution/state, and execution/stagedsync. Extended existing cleanup tests and confirmed their failures before adapting the merge.

+++ cl/persistence/format/snapshot_format/getters/execution_snapshot.go
@@ -86,10 +86,10 @@ func (r *ExecutionSnapshotReader) Withdrawals(number uint64, hash common.Hash) (
 	ret := solid.NewStaticListSSZ[*cltypes.Withdrawal](int(r.beaconCfg.MaxWithdrawalsPerPayload), 44)
 	for _, w := range body.Withdrawals {
 		ret.Append(&cltypes.Withdrawal{
-			Index:     w.Index,
-			Validator: w.Validator,
+			Index:     uint64(w.Index),
+			Validator: uint64(w.Validator),
 			Address:   w.Address,
-			Amount:    w.Amount,
+			Amount:    uint64(w.Amount),
 		})
 	}
 	return ret, nil
